Ajout du bouton pour l'affichage des catégories en mode mobile, du menu et correction des problèmes niveau responsive
