+++ src/app/components/Header.js
@@ -5,7 +5,7 @@ import NavBar from './navbar/NavBar';
 export default function Header() {
 
     return (
-        <header className="bg-white">
+        <header className="bg-white sticky-header">
             {/* Import de la Navbar */}
             <NavBar />
         </header>
