+++ src/app/page.tsx
@@ -2,6 +2,93 @@ export default function Home() {
   return (
     <div>
      <img src="./bannieres/home.webp" alt="" />
+     <br />
+      <h1>Home</h1>
+
+      <p>
+        Welcome to your new single-page application, built with:
+      </p>
+
+      <ul>
+        <li>
+          <a href="https://vitejs.dev/" target="_blank" rel="noopener noreferrer">Vite</a> and {' '}
+          <a href="https://reactjs.org" target="_blank" rel="noopener noreferrer">React</a>
+        </li>
+        <li>
+          <a href="https://reactrouter.com" target="_blank" rel="noopener noreferrer">React Router</a>
+        </li>
+        <li>
+          <a href="https://www.typescriptlang.org" target="_blank" rel="noopener noreferrer">TypeScript</a>
+        </li>
+
+        <li>
+          <a href="https://tailwindcss.com" target="_blank" rel="noopener noreferrer">Tailwind CSS</a>
+        </li>
+      </ul>
+      <ul>
+        <li>
+          <a href="https://vitejs.dev/" target="_blank" rel="noopener noreferrer">Vite</a> and {' '}
+          <a href="https://reactjs.org" target="_blank" rel="noopener noreferrer">React</a>
+        </li>
+        <li>
+          <a href="https://reactrouter.com" target="_blank" rel="noopener noreferrer">React Router</a>
+        </li>
+        <li>
+          <a href="https://www.typescriptlang.org" target="_blank" rel="noopener noreferrer">TypeScript</a>
+        </li>
+
+        <li>
+          <a href="https://tailwindcss.com" target="_blank" rel="noopener noreferrer">Tailwind CSS</a>
+        </li>
+      </ul>
+      <ul>
+        <li>
+          <a href="https://vitejs.dev/" target="_blank" rel="noopener noreferrer">Vite</a> and {' '}
+          <a href="https://reactjs.org" target="_blank" rel="noopener noreferrer">React</a>
+        </li>
+        <li>
+          <a href="https://reactrouter.com" target="_blank" rel="noopener noreferrer">React Router</a>
+        </li>
+        <li>
+          <a href="https://www.typescriptlang.org" target="_blank" rel="noopener noreferrer">TypeScript</a>
+        </li>
+
+        <li>
+          <a href="https://tailwindcss.com" target="_blank" rel="noopener noreferrer">Tailwind CSS</a>
+        </li>
+      </ul>
+      <ul>
+        <li>
+          <a href="https://vitejs.dev/" target="_blank" rel="noopener noreferrer">Vite</a> and {' '}
+          <a href="https://reactjs.org" target="_blank" rel="noopener noreferrer">React</a>
+        </li>
+        <li>
+          <a href="https://reactrouter.com" target="_blank" rel="noopener noreferrer">React Router</a>
+        </li>
+        <li>
+          <a href="https://www.typescriptlang.org" target="_blank" rel="noopener noreferrer">TypeScript</a>
+        </li>
+
+        <li>
+          <a href="https://tailwindcss.com" target="_blank" rel="noopener noreferrer">Tailwind CSS</a>
+        </li>
+      </ul>
+      <ul>
+        <li>
+          <a href="https://vitejs.dev/" target="_blank" rel="noopener noreferrer">Vite</a> and {' '}
+          <a href="https://reactjs.org" target="_blank" rel="noopener noreferrer">React</a>
+        </li>
+        <li>
+          <a href="https://reactrouter.com" target="_blank" rel="noopener noreferrer">React Router</a>
+        </li>
+        <li>
+          <a href="https://www.typescriptlang.org" target="_blank" rel="noopener noreferrer">TypeScript</a>
+        </li>
+
+        <li>
+          <a href="https://tailwindcss.com" target="_blank" rel="noopener noreferrer">Tailwind CSS</a>
+        </li>
+      </ul>
     </div>
   );
 }
